apiVersion, promises
- Adding `apiVersion()` method
- Now using ES6 promises over callbacks - unfortunately not able to get
ES7 async to compile with Webpack
- Moving to version 0.1.0 so I don’t have to keep updating Bower for
the demo whenever a new version is released
- Ad background is now semi-transparent black, which is visible when an
ad is smaller than the device. In the future, I might either create a
separate div under the ad for that background (so you don’t swipe that
background left/right with the ad) or allow ad managers to set that
background to match the ad color.