0.1.2
* “Campaigns” now known as “Publications,” using “ASID” to connect to
the API.
* User fingerprint sent to API server on `show()` to begin making ad
display smarter (it now does not display ads that have been swiped to
the right).
* Campaign ID returned from API server on `show()` so we can send it
back with `update()` on slide/tap to allow that ID to be logged in the
impression for improved analytics.
* Semi-transparent black background is now it’s own layer, so when the
ad is panned, that background stays in place.
* `util/util.js` imported into `hammer-as.js` to extend functionality
of Hammer functions.
* Removing `ua-parser-js` from Bower.JSON, as it’s now built into
Adswipe.JS.
* `util.findNextZIndex()` returns next highest z-index rather than just
storing in the config. This will let me create more layers rather than
everything being one layer.